+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - rename package from ubc to ubcpdk to match pypi name
 - move ubcsp into ubcpdk/simulations/circuits
+- rename ubcpdk/da as ubcpdk/data
 
 ## 0.0.12
